スコア計算用のクラスを分離
