+++ rulesets.xml
@@ -13,6 +13,7 @@
   <rule ref="rulesets/codesize.xml" />
   <rule ref="rulesets/cleancode.xml">
     <exclude name="ElseExpression" />
+    <exclude name="StaticAccess" />
   </rule>
   <rule ref="rulesets/controversial.xml" />
   <rule ref="rulesets/design.xml" />
